dont test creating of services by default

+++ bindings/java/src/net/hyperic/sigar/win32/test/TestService.java
@@ -5,12 +5,20 @@ import net.hyperic.sigar.win32.Service;
 
 public class TestService extends TestCase {
 
+    private static final boolean TEST_CREATE = false;
+    
     public TestService(String name) {
         super(name);
     }
 
-    public void testServiceCreateDelete() throws Exception {
+    public void testServiceOpen() throws Exception {
+        Service service = new Service("Eventlog");
+    }
 
+    public void testServiceCreateDelete() throws Exception {
+        if (!TEST_CREATE) {
+            return;
+        }
         Service service =
             Service.create("MyTestService", 
                            "My Test Service",
@@ -18,11 +26,10 @@ public class TestService extends TestCase {
                            "C:\\oracle\\ora90\\bin\\agntsrvc.exe");
     }
 
-    public void testServiceOpen() throws Exception {
-        Service service = new Service("MyTestService");
-    }
-
     public void testDeleteService() throws Exception {
+        if (!TEST_CREATE) {
+            return;
+        }
         Service service = new Service("MyTestService");
         service.delete();
     }
